Package each class into separate packages
