+++ app/src/main/java/com/amigo/course/Course.java
@@ -1,4 +1,4 @@
-package com.amigo.logic;
+package course;
 
 public class Course {
     private String courseCode;
