+++ app/src/test/java/com/amigo/app/AmigoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.amigo.amigo;
+package com.amigo.app;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
