+++ app/src/test/java/com/amigo/user/UserTextDatabaseTest.java
@@ -1,4 +1,4 @@
-package com.amigo.logic;
+package com.amigo.user;
 
 import org.junit.*;
 import static org.junit.Assert.*;
